Refactor storage engine helpers into named modules
