+++ src/storage/engines/utils/storageEngineHelpers.ts
@@ -1,6 +1,6 @@
-import { ParticipantData } from '../types';
-import type { CloudStorageEngine, StorageEngine } from './types';
-import { StudyConfig } from '../../parser/types';
+import { ParticipantData } from '../../types';
+import type { CloudStorageEngine, StorageEngine } from '../types';
+import { StudyConfig } from '../../../parser/types';
 
 export async function hash(input: string) {
   const msgUint8 = new TextEncoder().encode(input);
