+++ src/storage/tests/analysis-config.spec.ts
@@ -4,7 +4,7 @@ import {
 import { StudyConfig } from '../../parser/types';
 import testConfigSimple from './testConfigSimple.json';
 import testConfigSimple2 from './testConfigSimple2.json';
-import { hash } from '../engines/utils';
+import { hash } from '../engines/utils/storageEngineHelpers';
 import { buildConfigRows, ConfigInfo } from '../../analysis/individualStudy/config/utils';
 import { downloadConfigFile, downloadConfigFilesZip } from '../../utils/handleDownloadFiles';
 import { ConfigDiffModal } from '../../analysis/individualStudy/config/ConfigDiffModal';
